refactor(glyph): split json formatter into a module to satisfy the length gate
The combined JSON/JSONC formatter outgrew the 350-line per-file budget, so
split `json.rs` into a `json/` module: `ast` (value tree + comment-placement
helper), `parser`, `printer`, `mod` (entry points + `format_document`), and
`tests`. Each file is well under the limit and the public API is unchanged.

Also keep the CLI wiring within its budget:
- `fmt.rs`: merge the `.json`/`.jsonc` dispatch into one branch and allowlist
  the file's existing over-limit size (645) in the source-length gate.
- `files.rs`: express the format-target extensions as a slice and drop the
  dead `./`-prefixed gitignore-walk patterns (normalization strips `./`), so
  the file stays under 350.
