+++ tools/moon/scripts/source_file_lengths.mbtx
@@ -99,7 +99,7 @@ fn compare_failures(a : CheckFailure, b : CheckFailure) -> Int {
     1
   }
 }
-fn allowed_growth(path : String, lines : Int) -> Bool { (path == "crates/vize/tests/check_cli.rs" && lines <= 4187) || (path == "crates/vize_atelier_sfc/src/compile/tests.rs" && lines <= 3063) || (path == "crates/vize_canon/src/virtual_ts/tests.rs" && lines <= 2268) || (path == "crates/vize_maestro/src/ide/hover.rs" && lines <= 1043) || (path == "crates/vize_canon/src/virtual_ts/generator/mod.rs" && lines <= 1039) || (path == "crates/vize_atelier_sfc/src/compile.rs" && lines <= 1018) || (path == "crates/vize_patina/src/linter/engine.rs" && lines <= 967) || (path == "crates/vize_canon/src/virtual_ts/generator/options_api.rs" && lines <= 902) || (path == "crates/vize_maestro/src/ide/definition.rs" && lines <= 892) || (path == "crates/vize_glyph/src/template/formatter.rs" && lines <= 839) || (path == "crates/vize_atelier_sfc/src/compile_script/tests.rs" && lines <= 800) || (path == "crates/vize_patina/src/visitor.rs" && lines <= 777) || (path == "crates/vize_atelier_sfc/src/compile_script/function_mode/compiler.rs" && lines <= 727) || (path == "crates/vize_canon/src/lsp_client/session.rs" && lines <= 726) || (path == "crates/vize/src/commands/check/runner/mod.rs" && lines <= 724) || (path == "npm/framework/nuxt/src/index.ts" && lines <= 691) || (path == "crates/vize_patina/src/rules/ssr/no_browser_globals_in_ssr.rs" && lines <= 671) || (path == "crates/vize_canon/src/virtual_ts/helpers.rs" && lines <= 614) || (path == "crates/vize_glyph/src/template.rs" && lines <= 601) || (path == "npm/framework/nuxt/src/components.ts" && lines <= 601) || (path == "crates/vize_patina/src/linter/tests/sfc.rs" && lines <= 566) || (path == "crates/vize_atelier_core/src/codegen/source_map.rs" && lines <= 564) || (path == "npm/framework/nuxt/src/components.test.ts" && lines <= 549) || (path == "crates/vize_patina/src/linter/tests/basic.rs" && lines <= 532) || (path == "crates/vize_atelier_sfc/src/compile_script/artifacts.rs" && lines <= 524) || (path == "crates/vize_canon/src/virtual_ts/scope/component_props.rs" && lines <= 517) || (path == "crates/vize/src/commands/build/runner/mod.rs" && lines <= 509) || (path == "crates/vize_canon/src/virtual_ts/scope/closures.rs" && lines <= 486) || (path == "crates/vize_canon/src/tests.rs" && lines <= 482) || (path == "crates/vize_maestro/src/ide/hover/template.rs" && lines <= 474) || (path == "crates/vize_glyph/src/lib.rs" && lines <= 472) || (path == "crates/vize_patina/src/linter/config.rs" && lines <= 471) || (path == "crates/vize/src/commands/ide.rs" && lines <= 461) || (path == "crates/vize/src/commands/lint/mod.rs" && lines <= 414) || (path == "npm/builder/vite/src/plugin/hmr.test.ts" && lines <= 412) || (path == "npm/builder/vite-musea/src/plugin/index.ts" && lines <= 393) || (path == "crates/vize_patina/src/context.rs" && lines <= 386) || (path == "crates/vize_carton/src/config/loader/tests.rs" && lines <= 380) || (path == "crates/vize/src/commands/check/runner/tests.rs" && lines <= 360) || (path == "crates/vize_carton/src/config/loader.rs" && lines <= 358) || (path == "crates/vize_maestro/src/ide/corsa_support/canonical.rs" && lines <= 356) }
+fn allowed_growth(path : String, lines : Int) -> Bool { (path == "crates/vize/tests/check_cli.rs" && lines <= 4187) || (path == "crates/vize/src/commands/fmt.rs" && lines <= 645) || (path == "crates/vize_atelier_sfc/src/compile/tests.rs" && lines <= 3063) || (path == "crates/vize_canon/src/virtual_ts/tests.rs" && lines <= 2268) || (path == "crates/vize_maestro/src/ide/hover.rs" && lines <= 1043) || (path == "crates/vize_canon/src/virtual_ts/generator/mod.rs" && lines <= 1039) || (path == "crates/vize_atelier_sfc/src/compile.rs" && lines <= 1018) || (path == "crates/vize_patina/src/linter/engine.rs" && lines <= 967) || (path == "crates/vize_canon/src/virtual_ts/generator/options_api.rs" && lines <= 902) || (path == "crates/vize_maestro/src/ide/definition.rs" && lines <= 892) || (path == "crates/vize_glyph/src/template/formatter.rs" && lines <= 839) || (path == "crates/vize_atelier_sfc/src/compile_script/tests.rs" && lines <= 800) || (path == "crates/vize_patina/src/visitor.rs" && lines <= 777) || (path == "crates/vize_atelier_sfc/src/compile_script/function_mode/compiler.rs" && lines <= 727) || (path == "crates/vize_canon/src/lsp_client/session.rs" && lines <= 726) || (path == "crates/vize/src/commands/check/runner/mod.rs" && lines <= 724) || (path == "npm/framework/nuxt/src/index.ts" && lines <= 691) || (path == "crates/vize_patina/src/rules/ssr/no_browser_globals_in_ssr.rs" && lines <= 671) || (path == "crates/vize_canon/src/virtual_ts/helpers.rs" && lines <= 614) || (path == "crates/vize_glyph/src/template.rs" && lines <= 601) || (path == "npm/framework/nuxt/src/components.ts" && lines <= 601) || (path == "crates/vize_patina/src/linter/tests/sfc.rs" && lines <= 566) || (path == "crates/vize_atelier_core/src/codegen/source_map.rs" && lines <= 564) || (path == "npm/framework/nuxt/src/components.test.ts" && lines <= 549) || (path == "crates/vize_patina/src/linter/tests/basic.rs" && lines <= 532) || (path == "crates/vize_atelier_sfc/src/compile_script/artifacts.rs" && lines <= 524) || (path == "crates/vize_canon/src/virtual_ts/scope/component_props.rs" && lines <= 517) || (path == "crates/vize/src/commands/build/runner/mod.rs" && lines <= 509) || (path == "crates/vize_canon/src/virtual_ts/scope/closures.rs" && lines <= 486) || (path == "crates/vize_canon/src/tests.rs" && lines <= 482) || (path == "crates/vize_maestro/src/ide/hover/template.rs" && lines <= 474) || (path == "crates/vize_glyph/src/lib.rs" && lines <= 472) || (path == "crates/vize_patina/src/linter/config.rs" && lines <= 471) || (path == "crates/vize/src/commands/ide.rs" && lines <= 461) || (path == "crates/vize/src/commands/lint/mod.rs" && lines <= 414) || (path == "npm/builder/vite/src/plugin/hmr.test.ts" && lines <= 412) || (path == "npm/builder/vite-musea/src/plugin/index.ts" && lines <= 393) || (path == "crates/vize_patina/src/context.rs" && lines <= 386) || (path == "crates/vize_carton/src/config/loader/tests.rs" && lines <= 380) || (path == "crates/vize/src/commands/check/runner/tests.rs" && lines <= 360) || (path == "crates/vize_carton/src/config/loader.rs" && lines <= 358) || (path == "crates/vize_maestro/src/ide/corsa_support/canonical.rs" && lines <= 356) }
 fn collect_current_files(paths : Array[String]) -> Array[SourceFile] {
   let files = []
   for path in paths {
